+++ src/app.jsx
@@ -15,6 +15,7 @@ import CornerRibbon from './components/CornerRibbon';
 import dialogs from './components/dialogs';
 import Header from './components/header';
 import ServerConnectionManager from './components/ServerConnectionManager';
+import CollectiveRTHDialog from './features/collective-rth/CollectiveRTHDialog';
 import DetachedPanelManager from './features/detachable-panels/DetachedPanelManager';
 import DockDetailsDialog from './features/docks/DockDetailsDialog';
 import FirmwareUpdateSetupDialog from './features/firmware-update/FirmwareUpdateSetupDialog';
@@ -183,6 +184,7 @@ const App = ({ onFirstRender }) => (
       <SafetyDialog />
       <SavedLocationEditorDialog />
       <ShowConfiguratorDialog />
+      <CollectiveRTHDialog />
       <UAVDetailsDialog />
       <UploadDialog />
       <VersionCheckDialog />