+++ src/views/show-control/ShowConfiguratorButton.tsx
@@ -1,6 +1,7 @@
+import Home from '@mui/icons-material/Home';
+import IconButton from '@mui/material/IconButton';
 import ListItemButton from '@mui/material/ListItemButton';
 import ListItemText from '@mui/material/ListItemText';
-import type React from 'react';
 import { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { connect, useDispatch, useSelector } from 'react-redux';
@@ -13,28 +14,28 @@ import {
 } from '@skybrush/mui-components';
 
 import { Status } from '~/components/semantics';
+import { showDialog as showCollectiveRTHDialog } from '~/features/collective-rth/slice';
 import {
   isConnected,
   supportsStudioInterop,
 } from '~/features/servers/selectors';
+import { isDeveloperModeEnabled } from '~/features/session/selectors';
 import { showDialogAndClearUndoHistory } from '~/features/show-configurator/actions';
 import { selectShowConfiguratorDataFromShow } from '~/features/show-configurator/selectors';
-import { type ShowData } from '~/features/show-configurator/state';
+import { type ShowData } from '~/features/show-configurator/slice';
 import {
   getEnvironmentFromLoadedShowData,
   getOutdoorShowOrigin,
   getShowSegments,
   hasLoadedShowFile,
+  selectCollectiveRTHPlanSummary,
+  type CollectiveRTHPlanSummary,
 } from '~/features/show/selectors';
 import { getSetupStageStatuses } from '~/features/show/stages';
 import { showError } from '~/features/snackbar/actions';
-import { type PreparedI18nKey, tt } from '~/i18n';
+import { tt, type PreparedI18nKey } from '~/i18n';
 import Pro from '~/icons/Pro';
-import {
-  type AppDispatch,
-  type AppSelector,
-  type RootState,
-} from '~/store/reducers';
+import type { AppDispatch, AppSelector, RootState } from '~/store/reducers';
 import { type Nullable } from '~/utils/types';
 
 const PREREQUISITES: ReadonlyArray<
@@ -71,16 +72,26 @@ const PREREQUISITES: ReadonlyArray<
 
 type Props = Readonly<{
   base64Blob?: string;
+  devModeEnabled: boolean;
   show: ShowData | undefined;
   partialShow: Partial<ShowData>;
   // TODO: This should probably be a `ThunkActionDispatch`, but that doesn't
   //       seem to be reexported from `redux-thunk` via `@reduxjs/toolkit`...
   showDialogAndClearUndoHistory: (data?: ShowData) => void;
+  showCollectiveRTHDialog: () => void;
   status: Status;
+  rthPlanSummary: CollectiveRTHPlanSummary;
 }>;
 
-const ShowConfiguratorButton = (props: Props): React.JSX.Element => {
-  const { show, showDialogAndClearUndoHistory, status } = props;
+const ShowConfiguratorButton = (props: Props) => {
+  const {
+    devModeEnabled,
+    show,
+    showDialogAndClearUndoHistory,
+    showCollectiveRTHDialog,
+    status,
+    rthPlanSummary,
+  } = props;
 
   const dispatch = useDispatch<AppDispatch>();
   const { t } = useTranslation();
@@ -149,6 +160,36 @@ const ShowConfiguratorButton = (props: Props): React.JSX.Element => {
           }
           secondary={t('show.showConfigurator.description')}
         />
+        {devModeEnabled && (
+          <Tooltip
+            content={
+              rthPlanSummary.isValid
+                ? t('show.showConfigurator.tooltip.validRTHPlan', {
+                    numPlans: Object.keys(rthPlanSummary.plans).length,
+                  })
+                : t('show.showConfigurator.tooltip.invalidRTHPlan')
+            }
+            placement='left'
+          >
+            <IconButton
+              edge='end'
+              size='large'
+              color={
+                rthPlanSummary.numDrones === 0
+                  ? 'default'
+                  : rthPlanSummary.isValid
+                    ? 'success'
+                    : 'warning'
+              }
+              onClick={(evt) => {
+                evt.stopPropagation();
+                showCollectiveRTHDialog();
+              }}
+            >
+              <Home />
+            </IconButton>
+          </Tooltip>
+        )}
       </ListItemButton>
     </div>
   );
@@ -158,9 +199,12 @@ const ConnectedShowConfiguratorButton = connect(
   (state: RootState) => ({
     show: selectShowConfiguratorDataFromShow(state),
     status: getSetupStageStatuses(state).showConfigurator,
+    rthPlanSummary: selectCollectiveRTHPlanSummary(state),
+    devModeEnabled: isDeveloperModeEnabled(state),
   }),
   {
     showDialogAndClearUndoHistory,
+    showCollectiveRTHDialog,
   }
 )(ShowConfiguratorButton);